pyproject.toml
[project]
name = "vocso"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx==0.27.0",
    "huggingface-hub==0.20.1",
    "langchain>=0.1.20",
    "langchain-openai>=0.0.4",
    "numpy>=1.26.4",
    "openai==1.6.1",
    "pandas>=2.3.3",
    "pgvector==0.2.4",
    "psycopg2-binary==2.9.9",
    "python-dotenv>=1.2.1",
    "sentence-transformers==2.3.1",
    "sqlparse==0.4.4",
    "streamlit==1.29.0",
    "transformers==4.36.0",
]
